==> 4-analysis.Rmd <==
---
title: "EPLC Descriptives"
author: "Jacob W. Roden-Foreman"
date: "`r Sys.time()`"
output:
  html_document:
    pandoc_args: "--dpi=150"
    fig_height: 3
    fig_width: 5
    code_folding: show
    df_print: kable
    toc: true
    toc_depth: 3
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup and import}
knitr::opts_chunk$set(echo = TRUE,
											results = "asis")

library(tidyverse)
library(tidylog)
library(data.table)
library(fst)
library(bit64)
library(weights)

library(DiagrammeR)
library(DiagrammeRsvg)
library(rsvg)


mdata = read_fst("Data/final/final combined dataset.fst")
mdata$hosp_ed = as.factor(mdata$hosp_ed)


mdata$ecode1_label = paste0(mdata$ecode1,": ", mdata$ecode1_desc)




source("my_print.R")


med_iqi = function(x, w=mdata$discwt, na.rm=T, digits=0){
	
	digits = as.character(digits)
	
	fmt = paste0("%.", digits, "f [%.", digits, "f - %.", digits, "f]")
	
	if(is.null(w)){
		res = sprintf(fmt,
						quantile(x=x, probs = 0.5, na.rm = na.rm), 
						quantile(x=x, probs = 0.25, na.rm = na.rm), 
						quantile(x=x, probs = 0.75, na.rm = na.rm))
	} else{
		res = sprintf(fmt,
						spatstat::weighted.quantile(x=x, w=w, probs = 0.5, na.rm = na.rm), 
						spatstat::weighted.quantile(x=x, w=w, probs = 0.25, na.rm = na.rm), 
						spatstat::weighted.quantile(x=x, w=w, probs = 0.75, na.rm = na.rm))
	}
	
	#add thousand marks
	res = str_replace_all(res,
												"(\\d{1,})(\\d{3})",
												"\\1,\\2"
												)
	
	res
	
	
}



my.weighted.mean = function(x, w = mdata$discwt, na.rm=T){
	
	weighted.mean(x=x, w = w, na.rm = na.rm)
	
	
}


cat_perc = function(df, x, w = "discwt"){
	
	df = as.data.frame(df)
	x = df[, x]
	w = df[, w]
	rm(df)
	
	res = data.frame(var = unique(x))
	res$n = NA_real_
	res$perc = NA_real_
	total = sum(w, na.rm = T)
	
	for(i in 1:nrow(res)){
		res$n[i] = round( sum( w[x==res$var[i] ], na.rm = T ) )
	}
	
	res$var = as.character(res$var)
	res = res[!is.na(res$var), ]
	res = rbind(res, c("Missing", round(sum(w[ is.na(x) ])), NA))
	res$n = as.numeric(res$n)
	res$perc =  scales::percent((res$n/total), accuracy = 0.01)
	
	res$n = format(res$n, big.mark = ",")
	
	res
	
}



cat_perc_by = function(df, groups, wt_var = "discwt"){
	
	plyr::count(df, groups, wt_var = wt_var) %>% 
		dplyr::group_by_(groups[1]) %>% 
		dplyr::mutate(perc = scales::percent(freq/sum(freq), accuracy = 0.01)) %>% 
		dplyr::ungroup() %>% 
		dplyr::mutate(val = paste0(format(round(freq), big.mark = ","), " (", perc, ")")) %>% 
		dplyr::select(-freq, -perc) %>% 
		tidyr::pivot_wider(names_from = groups[1], 
											 values_from = val)
	
	
	
	
}




wilcox_weighted_p = function(y, x, w = mdata$discwt){
	y = data.table::frank(y)
	res = broom::glance(lm(y ~ x, weights = w))
	res = res[1, "p.value"]
	res = as.numeric(res)
	
	if(res < 0.0001){
		res = "< 0.0001"
	} else{
		res = round(res, 4)
	}
	
	res
	
}





```





\
\
\
\

# CONSORT Flow Chart
```{r}


pt_counts = read_lines("nobs log.txt") %>%
	str_split_fixed(., " = ",  n=2) %>% 
	as.data.frame() %>% 
	filter(str_detect(V1, "N_wt"))  #only want the weighted N


pt_counts$V2 = trimws(pt_counts$V2)


pt_counts$n = str_remove_all(pt_counts$V2, ",") %>% as.numeric()
pt_counts$n_change = lag(pt_counts$n, 1) - pt_counts$n
pt_counts$n_change = prettyNum(pt_counts$n_change, big.mark = ",")



pt_counts = rbind(pt_counts, 
									c("N_wt after all filters", 
										format(round(sum(mdata$discwt)), big.mark = ","),  
										round(sum(mdata$discwt)), NA ))



pt_counts$label = c("Weighted case count in NEDS dataset, 2010-2016",
										"Principal diagnosis of trauma",
										"Age \u2265 55 years",
										"Principal external cause of injury not one of burns, bites/stings, overexertion, poisoning, or misadventures of medical/surgical care",
										"Not missing data for mortality or total charges",
										"Not missing data for sex, ZIP code income quartile, total ED charges, or primary payer", 
										"Included for analysis"
) %>% str_wrap(., width = 35)



flow = grViz("
digraph graph2 {

graph [layout = dot, rankdir = TB]

# node definitions with substituted label text
node [shape = rectangle, width = 4, fillcolor = Biege]
a [label = '@@1']
b [label = '@@2']
c [label = '@@3']
d [label = '@@4']
e [label = '@@5']
f [label = '@@6']
g [label = '@@7']

b_out [label = '@@8']
c_out [label = '@@9']
d_out [label = '@@10']
e_out [label = '@@11']
f_out [label = '@@12']

{rank = same; b; b_out}
{rank = same; c; c_out}
{rank = same; d; d_out}
{rank = same; e; e_out}
{rank = same; f; f_out}



a -> b -> c -> d -> e -> f -> g
b -> b_out
c -> c_out 
d -> d_out
e -> e_out
f -> f_out

}

[1]: paste0(pt_counts$label[1], ': n = ', pt_counts$V2[1])
[2]: paste0(pt_counts$label[2], ': n = ', pt_counts$V2[2])
[3]: paste0(pt_counts$label[3], ': n = ', pt_counts$V2[3])
[4]: paste0(pt_counts$label[4], ': n = ', pt_counts$V2[4])
[5]: paste0(pt_counts$label[5], ': n = ', pt_counts$V2[5])
[6]: paste0(pt_counts$label[6], ': n = ', pt_counts$V2[6])
[7]: paste0(pt_counts$label[7], ': n = ', pt_counts$V2[7])
[8]: paste0('Excluded: n = ', pt_counts$n_change[2])
[9]: paste0('Excluded: n = ', pt_counts$n_change[3])
[10]: paste0('Excluded: n = ', pt_counts$n_change[4])
[11]: paste0('Excluded: n = ', pt_counts$n_change[5])
[12]: paste0('Excluded: n = ', pt_counts$n_change[6])
")

flow

flow %>% 
	export_svg() %>% 
	charToRaw() %>%
	rsvg_pdf("flowchart.pdf")
rm(flow)


```



\
\
\



# Table 1. Characteristics of `r format(sum(mdata$discwt), big.mark=",")` weighted cases with traumatic injury


```{r table 1}



count(mdata, high_risk) %>% 
	mutate(
				 n = paste0( scales::comma(round(n)), " (", scales::percent(n/sum(n), accuracy = 0.01), ")" )
	) %>% my_print()


count(mdata, high_risk, wt = discwt) %>% 
	mutate(
		n = paste0( scales::comma(round(n)), " (", scales::percent(n/sum(n), accuracy = 0.01), ")" )
	) %>% my_print()



tbl = mdata %>%
	group_by(high_risk) %>% 
	summarise(
		age = med_iqi(age, w = discwt),
		elixhauser = med_iqi(elixhauser, w = discwt),
		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
		los_ip = med_iqi(los_ip, w = discwt),
		totchg_ed = med_iqi(totchg_ed, w = discwt),
		totchg_ip_recode = med_iqi(totchg_ip_recode, w = discwt),
		totchg = med_iqi(totchg, w = discwt)
	) %>% t() %>% as.data.frame() 


ps = c(
	"p",
	wilcox_weighted_p(mdata$age, mdata$high_risk),
	wilcox_weighted_p(mdata$elixhauser, mdata$high_risk),
	wilcox_weighted_p(mdata$pDeath, mdata$high_risk),
	wilcox_weighted_p(mdata$los_ip, mdata$high_risk),
	wilcox_weighted_p(mdata$totchg_ed, mdata$high_risk),
	wilcox_weighted_p(mdata$totchg_ip_recode, mdata$high_risk),
	wilcox_weighted_p(mdata$totchg, mdata$high_risk)
)


cbind(rownames(tbl), tbl, ps) %>% my_print()




cat_perc_by(mdata, c("high_risk", "female")) %>% my_print()
weights::wtd.chi.sq(mdata$high_risk, mdata$female, 
										weight = mdata$discwt, mean1 = F) %>% knitr::kable()




tbl = bind_rows(
	cat_perc_by(mdata, c("high_risk", "injury_brain")) %>% filter(injury_brain), 	
	cat_perc_by(mdata, c("high_risk", "injury_skull_fx")) %>% filter(injury_skull_fx),
	cat_perc_by(mdata, c("high_risk", "injury_cspine")) %>% filter(injury_cspine),
	cat_perc_by(mdata, c("high_risk", "injury_rib_fx")) %>% filter(injury_rib_fx), 
	cat_perc_by(mdata, c("high_risk", "injury_cardio_pulm")) %>% filter(injury_cardio_pulm),
	cat_perc_by(mdata, c("high_risk", "injury_tspine")) %>% filter(injury_tspine),
	cat_perc_by(mdata, c("high_risk", "injury_lspine")) %>% filter(injury_lspine),
	cat_perc_by(mdata, c("high_risk", "injury_solid_abd")) %>% filter(injury_solid_abd),
	cat_perc_by(mdata, c("high_risk", "injury_hollow_abd")) %>% filter(injury_hollow_abd),
	cat_perc_by(mdata, c("high_risk", "injury_ue_fx")) %>% filter(injury_ue_fx),
	cat_perc_by(mdata, c("high_risk", "injury_pelvic_fx")) %>% filter(injury_pelvic_fx),
	cat_perc_by(mdata, c("high_risk", "injury_le_fx")) %>% filter(injury_le_fx) 
) %>% select("High-risk", "Non-high-risk")



ps = c(
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_brain, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_skull_fx, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_cspine, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_rib_fx, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_cardio_pulm, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_tspine, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_lspine, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_solid_abd, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_hollow_abd, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_ue_fx, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_pelvic_fx, 
											weight = mdata$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(mdata$high_risk, mdata$injury_le_fx, 
											weight = mdata$discwt, mean1 = F)["p.value"]
)

ps = Hmisc::format.pval(ps, eps = 0.0001, nsmall = 3)

cbind(tbl, ps) %>% my_print()




	


cat_perc_by(mdata, c("high_risk", "pay1_recode")) %>% my_print()
weights::wtd.chi.sq(mdata$high_risk, mdata$pay1_recode, 
										weight = mdata$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(mdata, c("high_risk", "zipinc_qrtl")) %>% my_print()
weights::wtd.chi.sq(mdata$high_risk, mdata$zipinc_qrtl, 
										weight = mdata$discwt, mean1 = F) %>% knitr::kable()

# cat_perc_by(mdata, c("high_risk", "died_visit")) %>% my_print()
# weights::wtd.chi.sq(mdata$high_risk, mdata$died_visit, 
# 										weight = mdata$discwt, mean1 = F) %>% knitr::kable()

# cat_perc_by(mdata, c("high_risk", "disp_ed")) %>% my_print()
# weights::wtd.chi.sq(mdata$high_risk, mdata$disp_ed, 
# 										weight = mdata$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(mdata, c("high_risk", "edevent")) %>% my_print()
weights::wtd.chi.sq(mdata$high_risk, mdata$edevent, 
										weight = mdata$discwt, mean1 = F) %>% knitr::kable()



ip = mdata %>% filter(disp_ed == "Admitted as inpatient") 
cat_perc_by(ip, c("high_risk", "disp_ip_recode")) %>% my_print()
weights::wtd.chi.sq(ip$high_risk, ip$disp_ip_recode, 
										weight = ip$discwt, mean1 = F) %>% knitr::kable()
rm(ip)


cat_perc_by(mdata, c("high_risk", "hosp_trauma_level")) %>% my_print()
weights::wtd.chi.sq(mdata$high_risk, mdata$hosp_trauma_level, 
										weight = mdata$discwt, mean1 = F) %>% knitr::kable()





```


```{r, sum-of-charges}

mdata %>% 
    summarise( scales::comma(sum(totchg * discwt)), 
               scales::comma(sum(discwt))
               )

mdata %>% 
	group_by(high_risk) %>% 
    summarise( scales::comma(sum(totchg * discwt)), 
               scales::comma(sum(discwt))
               )

```


```{r, LOS-of-lived-by-hrra}

temp = mdata %>%
	filter(edevent == "Admitted as inpatient") %>% 
	filter(mortality == 0)
	

temp %>% 
	group_by(high_risk) %>% 
	summarise(
		los_ip = med_iqi(los_ip, w = discwt)
		)  %>% 
	my_print()


wilcox_weighted_p(temp$los_ip, temp$high_risk, w = temp$discwt)



```


\
\
\
\



# Table 2. High-risk mechanisms of injury by TMPM probability of death, actual mortality, and median total charge  

## By high-risk groups  


```{r table 2}

mdata %>% 
	filter(high_risk == "High-risk") %>% 
	group_by(hr_group) %>% 
	summarise(
		n_obs = n(),
		n_weighted = round(sum(discwt)),
		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
		mort_rate = my.weighted.mean(mortality, w = discwt),
		
		death = paste0(scales::comma(sum(discwt[mortality==1]), accuracy = 1), " (", 
									 round(mort_rate*100, 2), "%)"),
		
		totchg = med_iqi(totchg, w = discwt)
	) %>% 
	mutate(obs_pct = n_obs/sum(n_obs),
				 obs = paste0(scales::comma(n_obs), " (", scales::percent(obs_pct, accuracy = 0.01), ")"),
				 wgt_pct = n_weighted/sum(n_weighted),
				 wgt = paste0(scales::comma(n_weighted), " (", scales::percent(wgt_pct, accuracy = 0.01), ")"),
				 #a bit ridiculous, but this gives descending rankings
				 tmpm_rank = n() - rank(pDeath) + 1,
				 mort_rank = rank(-mort_rate),
				 freq_rank = rank(-wgt_pct)
	) %>% 
	select(hr_group, 
				 # n_obs, obs_pct, 
				 # n_weighted, wgt_pct,
				 obs, wgt,
				 
				 pDeath, death, 
				 # mort_rate, 
				 totchg#, 
				 # tmpm_rank, mort_rank, freq_rank
	) %>% 
	my_print()

```

\
\
\
\


## By ecode group  

```{r, eval=FALSE}

mdata %>% 
	filter(high_risk == "High-risk") %>% 
	group_by(ecode1_group1) %>% 
	summarise(
		n_obs = n(),
		n_weighted = round(sum(discwt)),
		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
		mort_rate = my.weighted.mean(mortality, w = discwt),
		
		death = paste0(round(sum(discwt[mortality==1])), " (", 
									 round(mort_rate*100, 2), "%)"),
		
		totchg = med_iqi(totchg, w = discwt)
	) %>% 
	ungroup() %>% 
	mutate(obs_pct = n_obs/sum(n_obs),
				 wgt_pct = n_weighted/sum(n_weighted),
				 #a bit ridiculous, but this gives descending rankings
				 tmpm_rank = n() - rank(pDeath) + 1,
				 mort_rank = rank(-mort_rate),
				 freq_rank = rank(-wgt_pct)
	) %>% 
	select(ecode1_group1, 
				 n_obs, obs_pct, 
				 n_weighted, wgt_pct,
				 pDeath, death, mort_rate, totchg, 
				 tmpm_rank, mort_rank, freq_rank
	) %>% 
	my_print()







```

\
\
\
\


## By ecode subgroup  

```{r, eval=FALSE}


mdata %>% 
	filter(high_risk == "High-risk") %>% 
	group_by(ecode1_group2) %>% 
	summarise(
		n_obs = n(),
		n_weighted = round(sum(discwt)),
		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
		mort_rate = my.weighted.mean(mortality, w = discwt),
		
		death = paste0(round(sum(discwt[mortality==1])), " (", 
									 round(mort_rate*100, 2), "%)"),
		
		totchg = med_iqi(totchg, w = discwt)
	) %>% 
	ungroup() %>% 
	mutate(obs_pct = n_obs/sum(n_obs),
				 wgt_pct = n_weighted/sum(n_weighted),
				 #a bit ridiculous, but this gives descending rankings
				 tmpm_rank = n() - rank(pDeath) + 1,
				 mort_rank = rank(-mort_rate),
				 freq_rank = rank(-wgt_pct)
	) %>% 
	select(ecode1_group2, 
				 n_obs, obs_pct, 
				 n_weighted, wgt_pct,
				 pDeath, death, mort_rate, totchg, 
				 tmpm_rank, mort_rank, freq_rank
	) %>% 
	my_print()




```


\
\
\
\


## By ecode  


```{r}


mdata %>% 
	filter(high_risk == "High-risk") %>% 
	group_by(ecode1, ecode1_desc) %>% 
	summarise(
		n_obs = n(),
		n_weighted = round(sum(discwt)),
		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
		mort_rate = my.weighted.mean(mortality, w = discwt),
		
		death = paste0(round(sum(discwt[mortality==1])), " (", 
									 round(mort_rate*100, 2), "%)"),
		
		totchg = med_iqi(totchg, w = discwt)
	) %>% 
	ungroup() %>% 
	mutate(obs_pct = n_obs/sum(n_obs),
				 wgt_pct = n_weighted/sum(n_weighted),
				 #a bit ridiculous, but this gives descending rankings
				 tmpm_rank = n() - rank(pDeath) + 1,
				 mort_rank = rank(-mort_rate),
				 freq_rank = rank(-wgt_pct)
	) %>% 
	select(ecode1, ecode1_desc, 
				 n_obs, obs_pct, 
				 n_weighted, wgt_pct,
				 pDeath, death, mort_rate, totchg, 
				 tmpm_rank, mort_rank, freq_rank
	) %>% 
	my_print()

```



\
\
\
\

```{r}

# top_10 = mdata %>%
# 	filter(high_risk == "High-risk") %>%
# 	group_by(ecode1, ecode1_desc) %>%
# 	summarise(
# 		n_obs = n(),
# 		n_weighted = round(sum(discwt)),
# 		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
# 		totchg = med_iqi(totchg, w = discwt)
# 	) %>%
# 	ungroup() %>%
# 	arrange(desc(n_weighted)) %>%
# 	slice_head(n=10)
# 
# 
# 
# for(i in 1:10){
# 	ecode = top_10$ecode1[i] %>% as.character()
# 	flag_name = paste0("top_10_ecode_", ecode)
# 	mdata[, flag_name] = mdata$ecode1 == ecode
# }
# rm(ecode, flag_name, i)
# 
# 
# write_fst(mdata, "Data/final/final combined dataset with top 10.fst")



```


\
\
\
\

# Table 3. Characteristics of `r format(nrow(mdata[mdata$high_risk == "High-risk",]), big.mark=",")` encounters with traumatic injuries from high-risk activities by survival status  

```{r table 3}

mdata %>% 
	filter(high_risk == "High-risk") %>% 
	count(., mortality) %>% 
	mutate(pct = round((n/sum(n))*100,2),
				 n = format(n, big.mark = ","),
				 res = paste0(n, " (", pct, "%)")
	) %>% my_print()

mdata %>% 
	filter(high_risk == "High-risk") %>% 
	count(., mortality, wt = discwt) %>% 
	mutate(pct = round((n/sum(n))*100,2),
				 n = scales::comma(n),
				 res = paste0(n, " (", pct, "%)")
	) %>% my_print()




tbl = mdata %>% 
	filter(high_risk == "High-risk") %>% 
	group_by(mortality) %>% 
	summarise(
		age = med_iqi(age, w = discwt),
		elix = med_iqi(elixhauser, w = discwt),
		elix_max = max(elixhauser),
		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
		los_ip = med_iqi(los_ip, w = discwt),
		totchg_ed = med_iqi(totchg_ed, w = discwt),
		totchg_ip_recode = med_iqi(totchg_ip_recode, w = discwt),
		totchg = med_iqi(totchg, w = discwt)
	) %>% t() %>% as.data.frame() 


temp = mdata %>% filter(high_risk == "High-risk")



ps = c(
	"p",
	wilcox_weighted_p(temp$age, temp$mortality, temp$discwt),
	wilcox_weighted_p(temp$elixhauser, temp$mortality, temp$discwt),
	"NA",
	wilcox_weighted_p(temp$pDeath, temp$mortality, temp$discwt),
	wilcox_weighted_p(temp$los_ip, temp$mortality, temp$discwt),
	wilcox_weighted_p(temp$totchg_ed, temp$mortality, temp$discwt),
	wilcox_weighted_p(temp$totchg_ip_recode, temp$mortality, temp$discwt),
	wilcox_weighted_p(temp$totchg, temp$mortality, temp$discwt)
)


cbind(rownames(tbl), tbl, ps) %>% my_print()




temp %>% 
	group_by(hr_group, mortality) %>% 
	summarise(n = sum(discwt)) %>% 
	group_by(hr_group) %>% 
	mutate(
		pct = n/sum(n),
		res = paste0(scales::comma(n), " (", scales::percent(pct, 0.01), ")")
	) %>% 
	ungroup() %>% 
	select(hr_group, mortality, res) %>% 
	pivot_wider(
		id_cols = hr_group,
		names_from = mortality, 
		values_from = res
	) %>% 
	my_print()





weights::wtd.chi.sq(temp$mortality, temp$hr_group,
										weight = temp$discwt, mean1 = F) %>% knitr::kable()









cat_perc_by(temp, c("mortality", "female")) %>% my_print()
weights::wtd.chi.sq(temp$mortality, temp$female, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()



cat_perc_by(temp, c("mortality", "trauma_type")) %>% my_print()
weights::wtd.chi.sq(temp$mortality, temp$trauma_type, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()









tbl = bind_rows(
	cat_perc_by(temp, c("mortality", "injury_brain")) %>% filter(injury_brain), 	
	cat_perc_by(temp, c("mortality", "injury_skull_fx")) %>% filter(injury_skull_fx),
	cat_perc_by(temp, c("mortality", "injury_cspine")) %>% filter(injury_cspine),
	cat_perc_by(temp, c("mortality", "injury_rib_fx")) %>% filter(injury_rib_fx), 
	cat_perc_by(temp, c("mortality", "injury_cardio_pulm")) %>% filter(injury_cardio_pulm),
	cat_perc_by(temp, c("mortality", "injury_tspine")) %>% filter(injury_tspine),
	cat_perc_by(temp, c("mortality", "injury_lspine")) %>% filter(injury_lspine),
	cat_perc_by(temp, c("mortality", "injury_solid_abd")) %>% filter(injury_solid_abd),
	cat_perc_by(temp, c("mortality", "injury_hollow_abd")) %>% filter(injury_hollow_abd),
	cat_perc_by(temp, c("mortality", "injury_ue_fx")) %>% filter(injury_ue_fx),
	cat_perc_by(temp, c("mortality", "injury_pelvic_fx")) %>% filter(injury_pelvic_fx),
	cat_perc_by(temp, c("mortality", "injury_le_fx")) %>% filter(injury_le_fx) 
) %>% select("0", "1")



ps = c(
	weights::wtd.chi.sq(temp$mortality, temp$injury_brain, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_skull_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_cspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_rib_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_cardio_pulm, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_tspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_lspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_solid_abd, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_hollow_abd, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_ue_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_pelvic_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$mortality, temp$injury_le_fx, 
											weight = temp$discwt, mean1 = F)["p.value"]
)

ps = Hmisc::format.pval(ps, eps = 0.0001, nsmall = 3)

cbind(tbl, ps) %>% my_print()








cat_perc_by(temp, c("mortality", "pay1_recode")) %>% my_print()
weights::wtd.chi.sq(temp$mortality, temp$pay1_recode, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("mortality", "zipinc_qrtl")) %>% my_print()
weights::wtd.chi.sq(temp$mortality, temp$zipinc_qrtl, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("mortality", "disp_ed")) %>% my_print()
weights::wtd.chi.sq(temp$mortality, temp$disp_ed, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("mortality", "edevent")) %>% my_print()
weights::wtd.chi.sq(temp$mortality, temp$edevent, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()





ip = temp %>% filter(disp_ed == "Admitted as inpatient") 
cat_perc_by(ip, c("mortality", "disp_ip_recode")) %>% my_print()
weights::wtd.chi.sq(ip$mortality, ip$disp_ip_recode, 
										weight = ip$discwt, mean1 = F) %>% knitr::kable()
rm(ip)





cat_perc_by(temp, c("mortality", "hosp_trauma_level")) %>% my_print()
weights::wtd.chi.sq(temp$mortality, temp$hosp_trauma_level, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()



cat_perc_by(temp, c("mortality", "year_whole")) %>% my_print()
weights::wtd.chi.sq(temp$mortality, temp$year_whole, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()






```





\
\
\
\



# Table 3-2. Characteristics of `r format(nrow(mdata[mdata$high_risk == "High-risk",]), big.mark=",")` encounters with traumatic injuries from high-risk activities by age group and survival status  

```{r table 3-2}


temp = mdata %>% filter(high_risk == "High-risk")

temp$group = paste(temp$age_group, temp$mortality) %>% 
	str_replace("0", "lived") %>% 
	str_replace("1", "died") %>% 
	factor(levels = c("55-64 lived", "55-64 died",
										"65-84 lived", "65-84 died",
										"85+ lived", "85+ died"
										))


temp %>% 
	count(group) %>% 
	mutate(pct = scales::percent(n/sum(n)) ,
				 n = scales::comma(n),
				 res = paste0(n, " (", pct, ")")
	) %>% select(group, res) %>% t() %>% my_print()

temp %>% 
	count(group, wt = discwt) %>% 
	mutate(pct = scales::percent(n/sum(n)) ,
				 n = scales::comma(n),
				 res = paste0(n, " (", pct, ")")
	) %>% select(group, res) %>% t() %>% my_print()




tbl = temp %>% 
	group_by(group) %>% 
	summarise(
		age = med_iqi(age, w = discwt),
		elix = med_iqi(elixhauser, w = discwt),
		elix_max = max(elixhauser),
		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
		los_ip = med_iqi(los_ip, w = discwt),
		totchg_ed = med_iqi(totchg_ed, w = discwt),
		totchg_ip_recode = med_iqi(totchg_ip_recode, w = discwt),
		totchg = med_iqi(totchg, w = discwt)
	) %>% t() %>% as.data.frame() 



ps = c(
	"p",
	wilcox_weighted_p(temp$age, temp$group, temp$discwt),
	wilcox_weighted_p(temp$elixhauser, temp$group, temp$discwt),
	"NA",
	wilcox_weighted_p(temp$pDeath, temp$group, temp$discwt),
	wilcox_weighted_p(temp$los_ip, temp$group, temp$discwt),
	wilcox_weighted_p(temp$totchg_ed, temp$group, temp$discwt),
	wilcox_weighted_p(temp$totchg_ip_recode, temp$group, temp$discwt),
	wilcox_weighted_p(temp$totchg, temp$group, temp$discwt)
)


cbind(rownames(tbl), tbl, ps) %>% my_print()




temp %>% 
	group_by(hr_group, group) %>% 
	summarise(n = sum(discwt)) %>% 
	group_by(group) %>% 
	mutate(
		sum = sum(n),
		n = if_else(round(n)<=15, 15, n),
		pct = n/sum,
		res = paste0(scales::comma(n, 1), " (", scales::percent(pct, 0.01), ")")
	) %>% 
	ungroup() %>% 
	select(hr_group, group, res) %>% 
	pivot_wider(
		id_cols = hr_group,
		names_from = group, 
		values_from = res
	) %>% 
	my_print()





weights::wtd.chi.sq(temp$group, temp$hr_group,
										weight = temp$discwt, mean1 = F) %>% knitr::kable()









cat_perc_by(temp, c("group", "female")) %>% my_print()
weights::wtd.chi.sq(temp$group, temp$female, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()



cat_perc_by(temp, c("group", "trauma_type")) %>% my_print()
weights::wtd.chi.sq(temp$group, temp$trauma_type, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()









tbl = bind_rows(
	cat_perc_by(temp, c("group", "injury_brain")) %>% filter(injury_brain), 	
	cat_perc_by(temp, c("group", "injury_skull_fx")) %>% filter(injury_skull_fx),
	cat_perc_by(temp, c("group", "injury_cspine")) %>% filter(injury_cspine),
	cat_perc_by(temp, c("group", "injury_rib_fx")) %>% filter(injury_rib_fx), 
	cat_perc_by(temp, c("group", "injury_cardio_pulm")) %>% filter(injury_cardio_pulm),
	cat_perc_by(temp, c("group", "injury_tspine")) %>% filter(injury_tspine),
	cat_perc_by(temp, c("group", "injury_lspine")) %>% filter(injury_lspine),
	cat_perc_by(temp, c("group", "injury_solid_abd")) %>% filter(injury_solid_abd),
	cat_perc_by(temp, c("group", "injury_hollow_abd")) %>% filter(injury_hollow_abd),
	cat_perc_by(temp, c("group", "injury_ue_fx")) %>% filter(injury_ue_fx),
	cat_perc_by(temp, c("group", "injury_pelvic_fx")) %>% filter(injury_pelvic_fx),
	cat_perc_by(temp, c("group", "injury_le_fx")) %>% filter(injury_le_fx) 
)

tbl$var = c("injury_brain", "injury_skull_fx", "injury_cspine", 
						"injury_rib_fx", "injury_cardio_pulm", "injury_tspine",
						"injury_lspine", "injury_solid_abd", "injury_hollow_abd",
						"injury_ue_fx", "injury_pelvic_fx", "injury_le_fx")


tbl = tbl %>% select(var, contains("5"))




ps = c(
	weights::wtd.chi.sq(temp$group, temp$injury_brain, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_skull_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_cspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_rib_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_cardio_pulm, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_tspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_lspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_solid_abd, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_hollow_abd, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_ue_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_pelvic_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$group, temp$injury_le_fx, 
											weight = temp$discwt, mean1 = F)["p.value"]
)

ps = Hmisc::format.pval(ps, eps = 0.0001, nsmall = 3)

cbind(tbl, ps) %>% my_print()








cat_perc_by(temp, c("group", "pay1_recode")) %>% my_print()
weights::wtd.chi.sq(temp$group, temp$pay1_recode, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("group", "zipinc_qrtl")) %>% my_print()
weights::wtd.chi.sq(temp$group, temp$zipinc_qrtl, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("group", "disp_ed")) %>% my_print()
weights::wtd.chi.sq(temp$group, temp$disp_ed, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("group", "edevent")) %>% my_print()
weights::wtd.chi.sq(temp$group, temp$edevent, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()





ip = temp %>% filter(disp_ed == "Admitted as inpatient") 
cat_perc_by(ip, c("group", "disp_ip_recode")) %>% my_print()
weights::wtd.chi.sq(ip$group, ip$disp_ip_recode, 
										weight = ip$discwt, mean1 = F) %>% knitr::kable()
rm(ip)





cat_perc_by(temp, c("group", "hosp_trauma_level")) %>% my_print()
weights::wtd.chi.sq(temp$group, temp$hosp_trauma_level, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()



cat_perc_by(temp, c("group", "year_whole")) %>% my_print()
weights::wtd.chi.sq(temp$group, temp$year_whole, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()






```





\
\
\
\



# Table 4. Clinical profile of encounters injured in high-risk activities by age group  

```{r table 4}


temp %>% 
	filter(high_risk == "High-risk") %>% 
	count(age_group) %>% 
	mutate(pct = round((n/sum(n))*100, 2),
				 n = format(n, big.mark = ","),
				 res = paste0(n, " (", pct, "%)")
	) %>% t() %>% as.data.frame() %>% my_print()


temp %>% 
	filter(high_risk == "High-risk") %>% 
	count(age_group, wt = discwt) %>% 
	mutate(pct = round((n/sum(n))*100,2),
				 n = format(n, big.mark = ","),
				 res = paste0(n, " (", pct, "%)")
	) %>% t() %>% as.data.frame()  %>% my_print()



tbl = temp %>%
	group_by(age_group) %>% 
	summarise(
		age = med_iqi(age, w = discwt),
		elixhauser = med_iqi(elixhauser, w = discwt),
		pDeath = med_iqi(pDeath*100, w = discwt, digits=2),
		los_ip = med_iqi(los_ip, w = discwt),
		totchg_ed = med_iqi(totchg_ed, w = discwt),
		totchg_ip_recode = med_iqi(totchg_ip_recode, w = discwt),
		totchg = med_iqi(totchg, w = discwt)
	) %>% t() %>% as.data.frame() 


ps = c(
	"p",
	wilcox_weighted_p(temp$age, temp$age_group, temp$discwt),
	wilcox_weighted_p(temp$elixhauser, temp$age_group, temp$discwt),
	wilcox_weighted_p(temp$pDeath, temp$age_group, temp$discwt),
	wilcox_weighted_p(temp$los_ip, temp$age_group, temp$discwt),
	wilcox_weighted_p(temp$totchg_ed, temp$age_group, temp$discwt),
	wilcox_weighted_p(temp$totchg_ip_recode, temp$age_group, temp$discwt),
	wilcox_weighted_p(temp$totchg, temp$age_group, temp$discwt)
)


cbind(rownames(tbl), tbl, ps)  %>% my_print()






temp %>% 
	group_by(hr_group, age_group) %>% 
	summarise(n = sum(discwt)) %>% 
	group_by(hr_group) %>% 
	mutate(
		pct = n/sum(n),
		res = paste0(scales::comma(n), " (", scales::percent(pct, 0.01), ")")
	) %>% 
	ungroup() %>% 
	select(hr_group, age_group, res) %>% 
	pivot_wider(
		id_cols = hr_group,
		names_from = age_group, 
		values_from = res
	) %>% 
	my_print()





weights::wtd.chi.sq(temp$age_group, temp$hr_group,
										weight = temp$discwt, mean1 = F) %>% knitr::kable()









cat_perc_by(temp, c("age_group", "female")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$female, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()



cat_perc_by(temp, c("age_group", "trauma_type")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$trauma_type, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()









tbl = bind_rows(
	cat_perc_by(temp, c("age_group", "injury_brain")) %>% filter(injury_brain), 	
	cat_perc_by(temp, c("age_group", "injury_skull_fx")) %>% filter(injury_skull_fx),
	cat_perc_by(temp, c("age_group", "injury_cspine")) %>% filter(injury_cspine),
	cat_perc_by(temp, c("age_group", "injury_rib_fx")) %>% filter(injury_rib_fx), 
	cat_perc_by(temp, c("age_group", "injury_cardio_pulm")) %>% filter(injury_cardio_pulm),
	cat_perc_by(temp, c("age_group", "injury_tspine")) %>% filter(injury_tspine),
	cat_perc_by(temp, c("age_group", "injury_lspine")) %>% filter(injury_lspine),
	cat_perc_by(temp, c("age_group", "injury_solid_abd")) %>% filter(injury_solid_abd),
	cat_perc_by(temp, c("age_group", "injury_hollow_abd")) %>% filter(injury_hollow_abd),
	cat_perc_by(temp, c("age_group", "injury_ue_fx")) %>% filter(injury_ue_fx),
	cat_perc_by(temp, c("age_group", "injury_pelvic_fx")) %>% filter(injury_pelvic_fx),
	cat_perc_by(temp, c("age_group", "injury_le_fx")) %>% filter(injury_le_fx) 
) %>% select("55-64", "65-84", "85+")



ps = c(
	weights::wtd.chi.sq(temp$age_group, temp$injury_brain, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_skull_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_cspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_rib_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_cardio_pulm, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_tspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_lspine, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_solid_abd, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_hollow_abd, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_ue_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_pelvic_fx, 
											weight = temp$discwt, mean1 = F)["p.value"],
	weights::wtd.chi.sq(temp$age_group, temp$injury_le_fx, 
											weight = temp$discwt, mean1 = F)["p.value"]
)

ps = Hmisc::format.pval(ps, eps = 0.0001, nsmall = 3)

cbind(tbl, ps) %>% my_print()





cat_perc_by(temp, c("age_group", "pay1_recode")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$pay1_recode, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("age_group", "zipinc_qrtl")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$zipinc_qrtl, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("age_group", "disp_ed")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$disp_ed, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()

cat_perc_by(temp, c("age_group", "edevent")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$edevent, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()




ip = temp %>% filter(disp_ed == "Admitted as inpatient") 
cat_perc_by(ip, c("age_group", "disp_ip_recode")) %>% my_print()
weights::wtd.chi.sq(ip$age_group, ip$disp_ip_recode, 
										weight = ip$discwt, mean1 = F) %>% knitr::kable()
rm(ip)



cat_perc_by(temp, c("age_group", "hosp_trauma_level")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$hosp_trauma_level, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()



cat_perc_by(temp, c("age_group", "year_whole")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$year_whole, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()







cat_perc_by(temp, c("age_group", "mortality")) %>% my_print()
weights::wtd.chi.sq(temp$age_group, temp$mortality, 
										weight = temp$discwt, mean1 = F) %>% knitr::kable()












```



```{r, LOS-of-lived-by-age}


temp %>% 
	filter(mortality == 0) %>% 
	group_by(age_group) %>% 
	summarise(
		los_ip = med_iqi(los_ip, w = discwt)
		)  %>% 
	my_print()

with(
	temp %>% filter(mortality == 0), 
	wilcox_weighted_p(los_ip, age_group, w = discwt)
)




```


\
\
\
\






\
\
\
\





```{r} 
pander::pander(sessionInfo())
```
