==> 5-mortality-model.Rmd <==
---
title: "EPLC Mortality Model"
author: "Jacob W. Roden-Foreman"
date: "`r Sys.time()`"
output:
  html_document:
    pandoc_args: "--dpi=150"
    fig_height: 3
    fig_width: 5
    code_folding: show
    df_print: kable
    toc: true
    toc_depth: 3
    toc_float: true
editor_options: 
  chunk_output_type: console
---

# Setup and import  

```{r setup}
knitr::opts_chunk$set(echo = TRUE,
											results = "asis")

library(tidyverse)
library(tidylog)
library(data.table)
library(fst)
suppressMessages(library(bit64))
library(weights)
library(glmnet)
library(mgcv)
library(gammit)



extract_fixed = function(model, ci_level = 0.95, digits = 3){
	fe <- data.frame(summary(model, re.test=FALSE)$p.table)
	colnames(fe) = c("value", "se", "t", "p")
	if (ci_level > 0) {
		lower <- (1 - ci_level)/2
		upper <- 1 - lower
		nu <- model$df.residual
		mult <- stats::qt(upper, nu)
		ci <- data.frame(lower = fe$value - mult * fe$se, upper = fe$value + 
			mult * fe$se)
		colnames(ci) <- paste0(c("lower_", "upper_"), c(lower, 
			upper) * 100)
		fe <- data.frame(fe, ci)
	}
	fe <- fe %>% dplyr::mutate_all(round, digits = digits) %>% 
		dplyr::mutate(term = gsub(rownames(fe), pattern = "[\\(,\\)]", 
			replacement = "")) %>% dplyr::select(term, dplyr::everything()) %>% 
		dplyr::as_tibble()
	fe
}



```



\
\
\
\



# Preprocessing
```{r preprocessing, eval=FALSE}



mdata = read_fst("Data/final/final combined dataset.fst",
								 columns = c('hosp_ed', 'year_whole',
								 						'discwt', 'neds_stratum', 
								 						'age',
								 						'age_group',
								 						'female', 'zipinc_qrtl',
								 						'pay1_recode', 
								 						'pDeath', 'elixhauser',
								 						'mortality', 
								 						'high_risk', "hr_group", 
								 						'hosp_trauma_level',
								 						'injury_brain', 'injury_skull_fx',
								 						'injury_cspine', 'injury_rib_fx',
								 						'injury_cardio_pulm', 'injury_tspine',
								 						'injury_lspine', 'injury_solid_abd',
								 						'injury_hollow_abd', 'injury_ue_fx',
								 						'injury_pelvic_fx', 'injury_le_fx'
								 )
)

mdata = mdata %>% filter(high_risk == "High-risk")

mdata$hosp_ed = as.factor(mdata$hosp_ed)




#very small number of deaths in some groups
mdata %>% group_by(hr_group) %>% 
    summarise(n = sum(discwt),
              dead_n = sum(discwt[mortality==1]),
              mort_rate = scales::percent( dead_n / n, accuracy = 0.01)
    ) %>% 
    arrange(desc(dead_n))



#only use the four groups with >50 deaths
mdata = mdata %>% 
	filter(hr_group %in% c("Motorcycle Related", 
												 "Pedal Cycle Related", 
												 "Off-Road Vehicle", 
												 "Watercraft Related"))

mdata$hr_group = forcats::fct_drop(mdata$hr_group)





#separate everything out into (sparse) matrices to reduce memory pressure
Y = mdata$mortality

W = mdata$discwt


mdata = mdata %>% mutate_at(vars(starts_with("injury_")), as.integer)

mdata$zipinc_qrtl = factor(mdata$zipinc_qrtl, ordered = F)

library(Matrix)

temp = poly(mdata$age, 3)
mdata$age_L = temp[, 1]
mdata$age_Q = temp[, 2]
mdata$age_C = temp[, 3]
rm(temp)


mdata$pDeath_logit = arm::logit(mdata$pDeath)



X = sparse.model.matrix( ~ 
												 	#demographics
												 	age_L + age_Q + age_C + age_group + female + zipinc_qrtl + pay1_recode + 
												 	
												 	#case mix vars
												 	pDeath_logit + elixhauser +
												 	
												 	#high risk groups
												 	hr_group + 
												 	
												 	#injuries
												 	injury_brain + injury_skull_fx + injury_cspine + 
												 	injury_rib_fx + injury_cardio_pulm + injury_tspine + 
												 	injury_lspine + injury_solid_abd + injury_hollow_abd + 
												 	injury_ue_fx + injury_pelvic_fx + injury_le_fx + 
												 	
												 	#level 2 fixed effects
												 	hosp_trauma_level + year_whole
												 ,
												 mdata
)
X = X[, colnames(X) != "(Intercept)"] #remove the intercept










#define training (70%) and testing (30%) samples
set.seed(1111)
train = rbinom(n = nrow(mdata), size = 1, prob = 0.7) %>% as.logical()







```




# Variable Selection  

```{r vs, eval=FALSE}

#about 8 minutes
invisible(gc())
start = Sys.time()
start
cv_fit = cv.glmnet(y = Y[train], 
									 x = X[train, ],
									 weights = W[train],

									 #trace.it = 1, #print progress
									 
									 nfolds = 50,
									 family = "binomial", 
									 type.measure = "deviance", #use deviance as loss function
									 
									 #instead of default "Newton", which uses the exact hessian:
									 type.logistic = "modified.Newton", 
									 alpha = 1  #LASSO
									 
)
stop = Sys.time()
stop - start
beepr::beep()


plot(cv_fit)


# lambda.min is the value of lambda that gives minimum mean cross-validated error. 
log(cv_fit$lambda.min)
coef(cv_fit, s = "lambda.min") %>% as.matrix() %>% 
	data.frame() %>%  rownames_to_column() %>% 
	filter(X1 != 0) %>% arrange(rowname)
# has too many predictors



# lambda.1se is the most regularized model with loss that is within one standard error of model with the least loss.
log(cv_fit$lambda.1se)
coef(cv_fit, s = "lambda.1se") %>% as.matrix() %>% 
	data.frame() %>%  rownames_to_column() %>% 
	filter(X1 != 0) %>% arrange(rowname)
#takes everything out except age, age^2, age_group85+, 
# injury_cardio_pulm,
# injury_cspine, injury_hollow_abd, 
# injury_skull_fx, injury_ue_fx, 
# pDeath_logit, year_whole







#final choice:
coef(cv_fit, s = "lambda.1se") %>% as.matrix()

coefs = coef(cv_fit, s = "lambda.1se")
coefs = data.frame( var = rownames(coefs),  b = matrix(coefs)  )
coefs = coefs[coefs$b != 0, ]
coefs

rm(cv_fit, X, Y, W, start, stop)


```


\
\



# Testing Model

https://m-clark.github.io/posts/2019-10-20-big-mixed-models/#additive-models-as-mixed-models

## Model with HRRAs as random effects  


```{r testing-model-rand, eval=FALSE}
test = mdata %>% filter(train != TRUE)
format(nrow(test), big.mark = ",")
format(round(sum(test$discwt)), big.mark = ",")



test$hosp_trauma_level = factor(test$hosp_trauma_level,
																ordered = FALSE)

test$year_whole = factor(test$year_whole)

test = test %>% mutate_if(is.logical, as.integer)





#~8 mins 
start = Sys.time()
start
mort_mod = bam(mortality ~ 
							 	#level 1 fixed effects:
							 	poly(age, 3, raw = T) +  #keeping age^3 since LASSO included age, age^2, and age_group85+
							 	
							 	injury_skull_fx + injury_cspine +
							 	injury_cardio_pulm + injury_hollow_abd +
							 	injury_ue_fx + 
							 	
							 	
							 	pDeath_logit + elixhauser +  #keeping these two regardless of LASSO
							 	
							 	#level 2 fixed effects:
							 	hosp_trauma_level + year_whole + 
							 	
							 	#random effects via smoothers:
							 	s(hr_group, bs = "re") +
							 	s(hosp_ed, bs = "re") + 
							 	s(neds_stratum, bs = "re"),
							 
							 family = binomial("logit"),
							 weights = test$discwt,
							 data = test,
							 
							 discrete = TRUE,
							 control = list(nthreads = 3, trace = TRUE)
							 
)
end = Sys.time()
beepr::beep()

end - start

saveRDS(mort_mod, "Results/mort_mod_some_groups_randef.RDS")

options(show.signif.stars = F)

summary(mort_mod, re.test = F)


fixed = extract_fixed(mort_mod, ci_level = 1-0.005, digits = 5) 
colnames(fixed) = c("term", "value", "se", "t", "p", "lower_ci", "upper_ci")
fixed$OR = sprintf("%.2f", exp(fixed$value))
fixed$CI = paste0( sprintf("%.2f", exp(fixed$lower_ci)), ", ", 
									 sprintf("%.2f", exp(fixed$upper_ci)))
fixed$pval = if_else(fixed$p < 0.0001, "< 0.0001", sprintf("%.3f", fixed$p))
TCCD::my_print(fixed)





ranef = extract_ranef(mort_mod, re = "hr_group", ci_level = 1-0.005, digits = 5) 
colnames(ranef) = c("group_var", "effect", "group", "value", "se", "lower_ci", "upper_ci")
ranef$OR = sprintf("%.2f", exp(ranef$value))
ranef$CI = paste0( sprintf("%.2f", exp(ranef$lower_ci)), ", ", 
									 sprintf("%.2f", exp(ranef$upper_ci)))
ranef$pval = pnorm(abs(ranef$value/ranef$se) , lower.tail = F)*2
ranef$pval = if_else(ranef$pval < 0.0001, "< 0.0001", sprintf("%.3f", ranef$pval))
TCCD::my_print(ranef)

rm(mort_mod)
```





\
\


## Model with HRRAs as fixed effects  


```{r testing-model-fixed, eval=FALSE}
#adjust names for multiple comparisons later
test$hr_group = str_replace_all(test$hr_group, " ", "_")
test$hr_group = str_replace_all(test$hr_group, "-", "_")


#set Pedal Cycle Related as reference group (largest)
test$hr_group = factor(test$hr_group)
test$hr_group = relevel(test$hr_group, "Pedal_Cycle_Related")


#~8 mins 
start = Sys.time()
start
mort_mod2 = bam(mortality ~ 
									#level 1 fixed effects:
									
									hr_group +
									
									poly(age, 3, raw = T) +  #keeping age^3 since LASSO included age, age^2, and age_group85+
									
									injury_skull_fx + injury_cspine +
									injury_cardio_pulm + injury_hollow_abd +
									injury_ue_fx + 
									
									pDeath_logit + elixhauser +  #keeping these two regardless of LASSO
									
									
									#level 2 fixed effects:
									hosp_trauma_level + year_whole + 
									
									#random effects via smoothers:
									s(hosp_ed, bs = "re") + 
									s(neds_stratum, bs = "re"),
								
								family = binomial("logit"),
								weights = test$discwt,
								data = test,
								
								discrete = TRUE,
								control = list(nthreads = 3, trace = TRUE)
								
)
end = Sys.time()
beepr::beep()

end - start


saveRDS(mort_mod2, "Results/mort_mod_some_groups_fixed.RDS")


summary(mort_mod2, re.test = F)




fixed = extract_fixed(mort_mod2, ci_level = 1-0.005, digits = 5) 
colnames(fixed) = c("term", "value", "se", "t", "p", "lower_ci", "upper_ci")
fixed$OR = sprintf("%.2f", exp(fixed$value))
fixed$CI = paste0( sprintf("%.2f", exp(fixed$lower_ci)), ", ", 
									 sprintf("%.2f", exp(fixed$upper_ci)))
fixed$pval = if_else(fixed$p < 0.0001, "< 0.0001", sprintf("%.3f", fixed$p))
TCCD::my_print(fixed)






#need pairwise comparisons.  
# multcomp's function to automatically make contrasts (mcp) doesn't work with models from bam().
# solution from https://stats.stackexchange.com/questions/376237/correcting-for-multiple-pairwise-comparisons-with-gam-objects-mgcv-in-r


library(multcomp)




#comparisons
#Pedal_Cycle_Related is the ref. group in the model
pairs = test %>% distinct(hr_group) %>% pull() %>% 
	sort() %>% as.character() %>% 
	combn(2) %>% t() %>% data.frame()

library(glue)


#all comparisons against Pedal_Cycle_Related
comps = pairs %>% 
	filter(X1 == "Pedal_Cycle_Related") %>% 
	glue_data("hr_group{X2} = 0")



#all other comparisons
pairs = pairs %>% 
	filter(X1 != "Pedal_Cycle_Related") %>% 
	filter(X2 != "Pedal_Cycle_Related")


comps = c(comps, 
					pairs %>% glue_data("hr_group{X1} - hr_group{X2} = 0")
)


pw_comps = glht(mort_mod2, 
								linfct = comps
								)



summary(pw_comps, test = adjusted("holm") )


res = confint(pw_comps, level = 1-0.005, calpha = univariate_calpha())$confint

res = res %>% as.data.frame() %>% rownames_to_column()

res$ps = summary(pw_comps, test = adjusted("holm") )$test$pvalues

res$OR = scales::comma(exp(res$Estimate), accuracy = 0.01)
res$ci_l = scales::comma(exp(res$lwr), accuracy = 0.01)
res$ci_u = scales::comma(exp(res$upr), accuracy = 0.01)

res$OR_CI = paste0(res$OR, " (", res$ci_l, " - ", res$ci_u, ")")

res$pval = scales::pvalue(res$ps, accuracy = 0.0001)


res %>% dplyr::select(rowname, OR:pval) %>% TCCD::my_print()

rm(mort_mod2)

rm(coefs, fixed, pairs, pw_comps, ranef, res, test)

```




\
\
\


# Re-do with all high risk groups included

\

## Pre-process again  

```{r preprocessing2}


mdata = read_fst("Data/final/final combined dataset.fst",
								 columns = c('hosp_ed', 'year_whole',
								 						'discwt', 'neds_stratum', 
								 						'age',
								 						'age_group',
								 						'female', 'zipinc_qrtl',
								 						'pay1_recode', 
								 						'pDeath', 'elixhauser',
								 						'mortality', 
								 						'high_risk', "hr_group", 
								 						'hosp_trauma_level',
								 						'injury_brain', 'injury_skull_fx',
								 						'injury_cspine', 'injury_rib_fx',
								 						'injury_cardio_pulm', 'injury_tspine',
								 						'injury_lspine', 'injury_solid_abd',
								 						'injury_hollow_abd', 'injury_ue_fx',
								 						'injury_pelvic_fx', 'injury_le_fx'
								 )
)

mdata = mdata %>% filter(high_risk == "High-risk")

mdata$hosp_ed = as.factor(mdata$hosp_ed)







#separate everything out into (sparse) matrices to reduce memory pressure
Y = mdata$mortality

W = mdata$discwt


mdata = mdata %>% mutate_at(vars(starts_with("injury_")), as.integer)

mdata$zipinc_qrtl = factor(mdata$zipinc_qrtl, ordered = F)


mdata$age_capped = if_else(mdata$age > 89, 90L, mdata$age)


temp = poly(mdata$age_capped, 3)
mdata$age_capped_L = temp[, 1]
mdata$age_capped_Q = temp[, 2]
mdata$age_capped_C = temp[, 3]
rm(temp)



mdata$pDeath_logit = arm::logit(mdata$pDeath)




library(Matrix)

X = sparse.model.matrix( ~ 
												 	#demographics
												 	age_capped_L + age_capped_Q + age_capped_C + age_group + female + zipinc_qrtl + pay1_recode + 
												 	
												 	#case mix vars
												 	pDeath_logit + elixhauser +
												 	
												 	#high risk groups
												 	hr_group + 
												 	
												 	#injuries
												 	injury_brain + injury_skull_fx + injury_cspine + 
												 	injury_rib_fx + injury_cardio_pulm + injury_tspine + 
												 	injury_lspine + injury_solid_abd + injury_hollow_abd + 
												 	injury_ue_fx + injury_pelvic_fx + injury_le_fx + 
												 	
												 	#level 2 fixed effects
												 	hosp_trauma_level + year_whole
												 ,
												 mdata
)
X = X[, -1] #remove the intercept










#define training (70%) and testing (30%) samples
set.seed(1111)
train = rbinom(n = nrow(mdata), size = 1, prob = 0.7) %>% as.logical()





```


## Variable Selection again  

```{r vs2}

#about 10 minutes
invisible(gc())
start = Sys.time()
start
cv_fit = cv.glmnet(y = Y[train], 
									 x = X[train, ],
									 weights = W[train],

									 trace.it = interactive(), #print progress
									 
									 nfolds = 50,
									 family = "binomial", 
									 type.measure = "deviance", #use deviance as loss function
									 
									 #instead of default "Newton", which uses the exact hessian:
									 type.logistic = "modified.Newton", 
									 alpha = 1  #LASSO
									 
)
stop = Sys.time()
stop - start
beepr::beep()


plot(cv_fit)


# lambda.min is the value of lambda that gives minimum mean cross-validated error. 
log(cv_fit$lambda.min)
coef(cv_fit, s = "lambda.min") %>% as.matrix() %>% 
	data.frame() %>%  rownames_to_column() %>% 
	filter(X1 != 0) %>% arrange(rowname)
# has too many predictors



# lambda.1se is the most regularized model with loss that is within one standard error of model with the least loss.
log(cv_fit$lambda.1se)
coef(cv_fit, s = "lambda.1se") %>% as.matrix() %>% 
	data.frame() %>%  rownames_to_column() %>% 
	filter(X1 != 0) %>% arrange(rowname)
#takes everything out except age, age^2, age_group85+, 
# elixhauser, female, 
# injury_cspine, injury_hollow_abd, 
# injury_skull_fx, injury_ue_fx, 
# pay1_recodeSelf-pay, pDeath_logit, year_whole








#final choice:
coef(cv_fit, s = "lambda.1se") %>% as.matrix()

coefs = coef(cv_fit, s = "lambda.1se")
coefs = data.frame( var = rownames(coefs),  b = matrix(coefs)  )
coefs = coefs[coefs$b != 0, ]
coefs

rm(cv_fit, X, Y, W, start, stop)


```


\
\

## Testing Model 2

```{r}

test = mdata %>% filter(train != TRUE)
format(nrow(test), big.mark = ",")
format(round(sum(test$discwt)), big.mark = ",")



test$hosp_trauma_level = factor(test$hosp_trauma_level,
																ordered = FALSE)

test$year_whole = factor(test$year_whole)

test = test %>% mutate_if(is.logical, as.integer)



```



### Model with HRRAs as random effects  

```{r testing-model2-rand, eval=FALSE}


#~8 mins 
start = Sys.time()
start
mort_mod = bam(mortality ~ 
							 	#level 1 fixed effects:
							 	age_L + age_Q + age_C +  #keeping age^3 since LASSO included age, age^2, and age_group85+
							 	
							 	female + 
							 	
							 	pay1_recode + 
							 	
							 	injury_skull_fx + injury_cspine +
							 	injury_hollow_abd + injury_ue_fx + 
							 	
							 	pDeath_logit + elixhauser + #these two stay in a priori
							 	
							 	#level 2 fixed effects:
							 	hosp_trauma_level + year_whole + 
							 	
							 	#random effects via smoothers:
							 	s(hr_group, bs = "re") +
							 	s(hosp_ed, bs = "re") + 
							 	s(neds_stratum, bs = "re"),
							 
							 family = binomial("logit"),
							 weights = test$discwt,
							 data = test,
							 
							 discrete = TRUE,
							 control = list(nthreads = 3, trace = TRUE)
							 
)
end = Sys.time()
beepr::beep()

end - start

saveRDS(mort_mod, "Results/mort_mod_all_groups_ranef.RDS")

options(show.signif.stars = F)

summary(mort_mod, re.test = F)


fixed = extract_fixed(mort_mod, ci_level = 1-0.005, digits = 5) 
colnames(fixed) = c("term", "value", "se", "t", "p", "lower_ci", "upper_ci")
fixed$OR = sprintf("%.2f", exp(fixed$value))
fixed$CI = paste0( sprintf("%.2f", exp(fixed$lower_ci)), ", ", 
									 sprintf("%.2f", exp(fixed$upper_ci)))
fixed$pval = if_else(fixed$p < 0.0001, "< 0.0001", sprintf("%.3f", fixed$p))
TCCD::my_print(fixed)





ranef = extract_ranef(mort_mod, re = "hr_group", ci_level = 1-0.005, digits = 5) 
colnames(ranef) = c("group_var", "effect", "group", "value", "se", "lower_ci", "upper_ci")
ranef$OR = sprintf("%.2f", exp(ranef$value))
ranef$CI = paste0( sprintf("%.2f", exp(ranef$lower_ci)), ", ", 
									 sprintf("%.2f", exp(ranef$upper_ci)))
ranef$pval = pnorm(abs(ranef$value/ranef$se) , lower.tail = F)*2
ranef$pval = if_else(ranef$pval < 0.0001, "< 0.0001", sprintf("%.3f", ranef$pval))
TCCD::my_print(ranef)
```


\
\

### Model with HRRAs as fixed effects  


```{r testing-model2-fixed}

#adjust names for multiple comparisons later
test$hr_group = str_replace_all(test$hr_group, " ", "_")
test$hr_group = str_replace_all(test$hr_group, "-", "_")
test$hr_group = str_replace_all(test$hr_group, ",", "")
test$hr_group = str_replace_all(test$hr_group, "_&_", "_")
test$hr_group = str_replace_all(test$hr_group, "/", "_")



#set Pedal Cycle Related as reference group (largest)
test$hr_group = factor(test$hr_group)
test$hr_group = relevel(test$hr_group, "Pedal_Cycle_Related")



#center age_capped for linear interpretation
test$age_capped_cent70 = test$age_capped - 70



#~8 mins 
start = Sys.time()
start
mort_mod2 = bam(mortality ~ 
									#level 1 fixed effects:
									hr_group +
									
									age_capped_cent70 + I(age_capped_cent70^2) + #I(age^3) +  #keeping age^3 since LASSO included age, age^2, and age_group85+
									
									female + 
									
									# pay1_recode + 
									
									injury_skull_fx + injury_cspine +
									injury_hollow_abd + injury_ue_fx + 
									
									pDeath_logit + elixhauser + #these two stay in a priori
									
									
									#level 2 fixed effects:
									hosp_trauma_level + year_whole + 
									
									#random effects via smoothers:
									s(hosp_ed, bs = "re") + 
									s(neds_stratum, bs = "re"),
								
								family = binomial("logit"),
								weights = test$discwt,
								data = test,
								
								discrete = TRUE,
								control = list(nthreads = 3, trace = TRUE),
								drop.unused.levels = FALSE  #for predicting later
								
)
end = Sys.time()
beepr::beep()

end - start


saveRDS(mort_mod2, "Results/mort_mod_all_groups_fixed.RDS")


summary(mort_mod2, re.test = F)





extract_fixed = function(model, ci_level = 0.95, digits = 3){
	fe <- data.frame(summary(model, re.test=FALSE)$p.table)
	colnames(fe) = c("value", "se", "t", "p")
	if (ci_level > 0) {
		lower <- (1 - ci_level)/2
		upper <- 1 - lower
		nu <- model$df.residual
		mult <- stats::qt(upper, nu)
		ci <- data.frame(lower = fe$value - mult * fe$se, upper = fe$value + 
			mult * fe$se)
		colnames(ci) <- paste0(c("lower_", "upper_"), c(lower, 
			upper) * 100)
		fe <- data.frame(fe, ci)
	}
	fe <- fe %>% dplyr::mutate_all(round, digits = digits) %>% 
		dplyr::mutate(term = gsub(rownames(fe), pattern = "[\\(,\\)]", 
			replacement = "")) %>% dplyr::select(term, dplyr::everything()) %>% 
		dplyr::as_tibble()
	fe
}


fixed = extract_fixed(mort_mod2, ci_level = 1-0.005, digits = 5) 
colnames(fixed) = c("term", "value", "se", "t", "p", "lower_ci", "upper_ci")
fixed$OR = sprintf("%.2f", exp(fixed$value))
fixed$CI = paste0( sprintf("%.2f", exp(fixed$lower_ci)), ", ", 
									 sprintf("%.2f", exp(fixed$upper_ci)))
fixed$pval = if_else(fixed$p < 0.0001, "< 0.0001", sprintf("%.3f", fixed$p))
TCCD::my_print(fixed)







#look into age association some

temp = test %>% 
	group_by(age_capped) %>% 
	summarise(
		n = sum(discwt), 
		mort = sum(discwt*mortality)/n
	)

plot(temp$age_capped, temp$mort, type = "b"); abline(h = 0, col = "red")

p = sjPlot::plot_model(mort_mod2, type = "pred", term = "age_capped_cent70", ci.lvl = 1-0.005)

p = p + 
	scale_y_continuous(
		limits = c(0, NA), 
		labels = scales::percent_format(0.01),
		expand = expansion(c(0, 0.05)),
		name = "Predicted mortality rate"
	) +
	scale_x_continuous(
		breaks = seq(-15, 20, 5),
		labels = seq(55, 90, 5),
		name = "Age"
	) +
	theme_bw() +
	labs(title = NULL)

p

p$data %>% TCCD::my_print()
attr(p$data, "constant.values")



ggsave("Results/model predicted mortality by age at refs.pdf", 
			 plot = p,
			 device = "pdf", 
			 height = 4, width = 5)





library(visibly)

p2 = plot_gam(mort_mod2,
         conditional_data = tibble(
         	age_capped_cent70 = seq(-15, 20, 1),
         	hr_group = "Motorcycle_Related", 
         	female = mean(test$female), 
         	injury_skull_fx = mean(test$injury_skull_fx), 
         	injury_cspine = mean(test$injury_cspine), 
         	injury_hollow_abd = mean(test$injury_hollow_abd), 
         	injury_ue_fx = mean(test$injury_ue_fx), 
         	pDeath_logit = mean(test$pDeath_logit), 
         	elixhauser = mean(test$elixhauser), 
         	hosp_trauma_level = "Trauma center level I" #, 
         	# year_whole = , 
         	# hosp_ed = , 
         	# neds_stratum = 
         		),
							main_var = age_capped_cent70,
							line_color = "black", 
							ribbon_color = alpha("white", 0))


p2$data %>% TCCD::my_print()


p2 = p2 + 
	scale_y_continuous(
		limits = c(0, 0.01), 
		labels = scales::percent_format(0.01),
		expand = expansion(c(0, 0.05)),
		name = "Predicted mortality rate"
	) +
	scale_x_continuous(
		breaks = seq(-15, 20, 5),
		labels = seq(55, 90, 5),
		name = "Age"
	) +
	theme_bw() +
	labs(title = NULL)

cap = "\nEstimated marginal means calculated with continuous covariates set to their mean values, binary covariates set to their mean proportions, high-risk recreational activity group set to 'Motorcycle Related', and other categorical covariates set to their reference levels."

p2 = p2 + labs(caption = str_wrap(cap, 70)) + 
	theme(plot.caption = element_text(size = rel(0.6)))

p2


ggsave("Results/model predicted mortality by age at means.pdf", 
			 plot = p2,
			 device = "pdf", 
			 height = 5, width = 5)








#add CI and observed mortality

#based on visibly:::plot_gam_1d, which is normally called by visibly:::plot_gam
alpha = 0.005

conditional_data = tibble(
	age_capped_cent70 = test$age_capped_cent70,  #seq(-15, 20, 1),
	hr_group = "Motorcycle_Related", 
	female = mean(test$female), 
	injury_skull_fx = mean(test$injury_skull_fx), 
	injury_cspine = mean(test$injury_cspine), 
	injury_hollow_abd = mean(test$injury_hollow_abd), 
	injury_ue_fx = mean(test$injury_ue_fx), 
	pDeath_logit = mean(test$pDeath_logit), 
	elixhauser = mean(test$elixhauser), 
	hosp_trauma_level = "Trauma center level I",
	year_whole = test$year_whole,
	hosp_ed = test$hosp_ed,
	neds_stratum = test$neds_stratum
) 

conditional_data = conditional_data %>%
	bind_cols(tibble::as_tibble(predict_gamm(mort_mod2, ., re_form = NA, se = TRUE))) %>% 
	mutate(
		ll = mort_mod2$family$linkinv(prediction - qnorm(alpha/2, lower.tail = F) * se), 
		ul = mort_mod2$family$linkinv(prediction + qnorm(alpha/2, lower.tail = F) * se), 
		fit = mort_mod2$family$linkinv(prediction),
		age_capped = age_capped_cent70 + 70
	)




p3 = conditional_data %>% 
	group_by(age_capped) %>% 
	summarise(
		fit = mean(fit),
		ll = mean(ll),
		ul = mean(ul)
	) %>% 
	ungroup() %>% 
	mutate(
		fit_l = loess(fit ~ age_capped, data = .) %>% fitted.values(),
		ll_l = loess(ll ~ age_capped, data = .) %>% fitted.values(),
		ul_l = loess(ul ~ age_capped, data = .) %>% fitted.values()
	) %>% 
	ggplot(aes(x = age_capped, y = fit_l)) + 
	geom_ribbon(aes(ymin = ll_l, ymax = ul_l), fill = alpha("gray", 0.5)) + 
	geom_line(color = "black") 




#add observed mortality
om = mdata %>% 
	group_by(age_capped) %>% 
	summarise(
		mort = mean(mortality)
	)
	

p3 = p3 + 
	geom_smooth(aes(x = age_capped, y = mort), data = om, se=F,
							color = "black", linetype = "dashed", size = 0.8
							)


p3$data %>% TCCD::my_print()



p3 = p3 + 
	scale_y_continuous(
		limits = c(0, NA), 
		labels = scales::percent_format(1),
		expand = expansion(c(0, 0.05)),
		name = "Predicted mortality rate"
	) +
	scale_x_continuous(
		breaks = seq(55, 90, 5),
		labels = seq(55, 90, 5),
		name = "Age"
	) +
	theme_bw() +
	labs(title = NULL)


# cap = "\nEstimated marginal means calculated with continuous covariates set to their mean values, binary covariates set to their mean proportions, high-risk recreational activity group set to 'Motorcycle Related', and other categorical covariates set to their reference levels; random effects were not used."
# 
# p3 = p3 + labs(caption = str_wrap(cap, 110)) + 
# 	theme(plot.caption = element_text(size = rel(0.6)))

p3


ggsave("Results/model predicted mortality with CI by age at means.pdf", 
			 plot = p3,
			 device = "pdf", 
			 height = 5, width = 5)

	

rm(conditional_data, om)











#need pairwise comparisons.  
# multcomp's function to automatically make contrasts (mcp) doesn't work with models from bam().
# solution from https://stats.stackexchange.com/questions/376237/correcting-for-multiple-pairwise-comparisons-with-gam-objects-mgcv-in-r


library(multcomp)


#comparisons
#Pedal_Cycle_Related is the ref. group in the model
pairs = test %>% distinct(hr_group) %>% pull() %>% 
	sort() %>% as.character() %>% 
	combn(2) %>% t() %>% data.frame()

library(glue)


#all comparisons against Pedal_Cycle_Related
comps = pairs %>% 
	filter(X1 == "Pedal_Cycle_Related") %>% 
	glue_data("hr_group{X2} = 0")



#all other comparisons
pairs = pairs %>% 
	filter(X1 != "Pedal_Cycle_Related") %>% 
	filter(X2 != "Pedal_Cycle_Related")


comps = c(comps, 
					pairs %>% glue_data("hr_group{X1} - hr_group{X2} = 0")
)




pw_comps = glht(mort_mod2, 
								linfct = comps
								)



summary(pw_comps, test = adjusted("holm") )


res = confint(pw_comps, level = 1-0.005, calpha = univariate_calpha())$confint

res = res %>% as.data.frame() %>% rownames_to_column()

res$ps = summary(pw_comps, test = adjusted("none") )$test$pvalues
res$ps_adj = summary(pw_comps, test = adjusted("holm") )$test$pvalues

res$OR = scales::comma(exp(res$Estimate), accuracy = 0.01)
res$ci_l = scales::comma(exp(res$lwr), accuracy = 0.01)
res$ci_u = scales::comma(exp(res$upr), accuracy = 0.01)

res$OR_CI = paste0(res$OR, " (", res$ci_l, " - ", res$ci_u, ")")

res$ps = scales::pvalue(res$ps, accuracy = 0.0001)
res$ps_adj = scales::pvalue(res$ps_adj, accuracy = 0.0001)


res %>% dplyr::select(rowname, OR:OR_CI, ps, ps_adj) %>% TCCD::my_print()

rm(mort_mod, mort_mod2)

```




\
\
\
\





```{r} 
pander::pander(sessionInfo())
```
