==> 6-charges-model.Rmd <==
---
title: "EPLC Total Charges Model"
author: "Jacob W. Roden-Foreman"
date: "`r Sys.time()`"
output:
  html_document:
    pandoc_args: "--dpi=150"
    fig_height: 3
    fig_width: 5
    code_folding: show
    df_print: kable
    toc: true
    toc_depth: 3
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup}
knitr::opts_chunk$set(echo = TRUE,
                      results = "asis")

library(tidyverse)
suppressMessages(library(tidylog))
library(data.table)
library(fst)
suppressMessages(library(bit64))
library(weights)
library(glmnet)
library(mgcv)
library(gammit)


extract_fixed = function(model, ci_level = 0.95, digits = 3){
	fe <- data.frame(summary(model, re.test=FALSE)$p.table)
	colnames(fe) = c("value", "se", "t", "p")
	if (ci_level > 0) {
		lower <- (1 - ci_level)/2
		upper <- 1 - lower
		nu <- model$df.residual
		mult <- stats::qt(upper, nu)
		ci <- data.frame(lower = fe$value - mult * fe$se, upper = fe$value + 
			mult * fe$se)
		colnames(ci) <- paste0(c("lower_", "upper_"), c(lower, 
			upper) * 100)
		fe <- data.frame(fe, ci)
	}
	fe <- fe %>% dplyr::mutate_all(round, digits = digits) %>% 
		dplyr::mutate(term = gsub(rownames(fe), pattern = "[\\(,\\)]", 
			replacement = "")) %>% dplyr::select(term, dplyr::everything()) %>% 
		dplyr::as_tibble()
	fe
}



```


\
\
\

# Distribution of total charges

```{r estimate dist}



mdata = read_fst("Data/final/final combined dataset.fst", 
                 columns = c("totchg", "los_ip", "edevent", "discwt", "high_risk", "npr_ip")
)


mdata = mdata %>%
  filter(high_risk == "High-risk") %>% 
  select(-high_risk)



summary(mdata$totchg)

format(quantile(mdata$totchg, probs = seq(0, 1, 0.05)), big.mark = ",")

format(quantile(mdata$totchg, probs = seq(0.9, 1, 0.01)), big.mark = ",")

format(quantile(mdata$totchg, probs = seq(0.99, 1, 0.001)), big.mark = ",")

format(quantile(mdata$totchg, probs = seq(0.999, 1, 0.0001)), big.mark = ",")





counts = mdata %>% count(totchg) %>% 
  mutate(pct = n/sum(n), 
         cumpct = cumsum(pct)) %>% 
  mutate_if(is.numeric, round, 4)


counts %>% 
  ggplot(aes(x = totchg, y = n)) + 
  geom_col() +
  scale_x_binned(n.breaks = 50, labels = scales::comma) +
  theme_bw() + 
  theme(panel.grid.major.x = element_blank(),
        axis.text.x = element_text(angle = -90))




#Winsorize values greater than the 99th percentile
cut = quantile(mdata$totchg, probs = 0.99)
cut
mdata$totchg_winsor = mdata$totchg
mdata$totchg_winsor[mdata$totchg > cut] = cut+1



mdata %>% count(totchg_winsor) %>% 
  mutate(pct = n/sum(n), 
         cumpct = cumsum(pct)) %>% 
  mutate_if(is.numeric, round, 4) %>% 
  ggplot(aes(x = totchg_winsor, y = n)) + 
  geom_col() +
  scale_x_binned(n.breaks = 50, labels = scales::comma) +
  theme_bw() + 
  theme(panel.grid.major.x = element_blank(),
        axis.text.x = element_text(angle = -90))










#look at Poisson, NB, and gamma

#at the least, deviance residuals should be normally distributed
#optimally, deviance residuals should not be heteroskedastic, but that's not a big real for now
# because these models won't account for clustering


mdata$los_ip[which(mdata$edevent != "Admitted as inpatient")] = 0

mdata$npr_ip[which(mdata$edevent != "Admitted as inpatient")] = 0


par(mfrow=c(1, 2))

poismod = glm(totchg_winsor ~ edevent * los_ip + npr_ip, data = mdata,
              family = poisson("log"), weights = mdata$discwt)
ypred = predict(poismod, type = "link")
d_resid = residuals(poismod, type = 'deviance')
fam = poismod$family$family
link = poismod$family$link
hist(d_resid, breaks = 100, main = paste(fam, "with", link, "link"))
smoothScatter(x=ypred, y=d_resid, main = paste(fam, "with", link, "link"), nbin=800)
rm(ypred, d_resid, fam, link)



nbmod = MASS::glm.nb(totchg_winsor ~ edevent * los_ip + npr_ip, data = mdata,
                     weights = mdata$discwt)
ypred = predict(nbmod, type = "link")
d_resid = residuals(nbmod, type = 'deviance')
fam = nbmod$family$family
link = nbmod$family$link
hist(d_resid, breaks = 100, main = paste(fam, "with", link, "link"))
smoothScatter(x=ypred, y=d_resid, main = paste(fam, "with", link, "link"), nbin=800)
rm(ypred, d_resid, fam, link)




gammod = glm(totchg_winsor ~ edevent * los_ip + npr_ip, data = mdata,
             family = Gamma("log"), weights = mdata$discwt)
ypred = predict(gammod, type = "link")
d_resid = residuals(gammod, type = 'deviance')
fam = gammod$family$family
link = gammod$family$link
hist(d_resid, breaks = 100, main = paste(fam, "with", link, "link"))
smoothScatter(x=ypred, y=d_resid, main = paste(fam, "with", link, "link"), nbin=800)
rm(ypred, d_resid, fam, link)

par(mfrow = c(1, 1))







rm(poismod, nbmod, gammod, counts)

```


\
\




# Preprocessing
```{r preprocessing}


mdata = read_fst("Data/final/final combined dataset.fst",
                 columns = c("totchg",
                            
                 						'hosp_ed', 'year_whole',
								 						'discwt', 'neds_stratum',  
								 						'hosp_trauma_level',
                 						
                             #demos
                             "age", "age_group", 'female', 'zipinc_qrtl', 'pay1_recode', 
								 						'high_risk', "hr_group",
                            
								 						'pDeath', 'elixhauser', 
								 						
                             #outcomes
                             'edevent', 'disp_ip_recode', 
                             'los_ip', 'npr_ip', 
                             
                             #injuries
                             'injury_brain', 'injury_skull_fx',
								 						'injury_cspine', 'injury_rib_fx',
								 						'injury_cardio_pulm', 'injury_tspine',
								 						'injury_lspine', 'injury_solid_abd',
								 						'injury_hollow_abd', 'injury_ue_fx',
								 						'injury_pelvic_fx', 'injury_le_fx')
)



mdata = mdata %>% filter(high_risk == "High-risk") %>% select(-high_risk)





#Winsorize values greater than the 99th percentile
cut = quantile(mdata$totchg, probs = 0.99)
cut
mdata$totchg_winsor = mdata$totchg
mdata$totchg_winsor[mdata$totchg > cut] = cut+1




#if these are missing and pt was not admitted, set to zero
mdata$los_ip[which(is.na(mdata$los_ip) & mdata$edevent != "Admitted as inpatient" )] = 0

mdata$npr_ip[which(is.na(mdata$npr_ip) & mdata$edevent != "Admitted as inpatient")] = 0






#separate everything out into matrices to reduce memory pressure
Y = mdata$totchg_winsor

W = mdata$discwt


mdata = mdata %>% mutate_at(vars(starts_with("injury_")), as.integer)

mdata$age_capped = if_else(mdata$age > 89, 90L, mdata$age)

temp = poly(mdata$age_capped, 3)
mdata$age_capped_L = temp[, 1]
mdata$age_capped_Q = temp[, 2]
mdata$age_capped_C = temp[, 3]
rm(temp)


# temp = poly(mdata$los_ip, 3)
# mdata$los_ip_L = temp[, 1]
# mdata$los_ip_Q = temp[, 2]
# mdata$los_ip_C = temp[, 3]
# rm(temp)
# 
# 
# temp = poly(mdata$npr_ip, 3)
# mdata$npr_ip_L = temp[, 1]
# mdata$npr_ip_Q = temp[, 2]
# mdata$npr_ip_C = temp[, 3]
# rm(temp)




#set reference levels for a few vars
mdata$edevent = factor(mdata$edevent) %>% 
  relevel(ref = "Treated and released")


mdata$disp_ip_recode = factor(mdata$disp_ip_recode) %>% 
  relevel(ref = "Not admitted as inpatient")

mdata$zipinc_qrtl = factor(mdata$zipinc_qrtl, ordered = F)



X = model.matrix( ~ 
                    #demographics
                    age_capped_L + age_capped_Q + age_capped_C + 
										age_group + female + zipinc_qrtl + pay1_recode + 
										
										#case mix vars
										pDeath + elixhauser +
										
										#high risk groups
										hr_group + 
										
                    #injuries
                    injury_brain + injury_skull_fx + injury_cspine + 
                    injury_rib_fx + injury_cardio_pulm + injury_tspine + 
                    injury_lspine + injury_solid_abd + injury_hollow_abd + 
                    injury_ue_fx + injury_pelvic_fx + injury_le_fx +
                    
                    #outcomes
                    edevent + disp_ip_recode + 
                    
                    #including quadratic and cubic effects sucks up all the variance
                    los_ip + #los_ip_Q + los_ip_C +
                    npr_ip + #npr_ip_Q + #npr_ip_C 
										
										
										
										#clustering vars
										hosp_trauma_level + year_whole
										
                    ,
                  mdata
)

X = X[, colnames(X) != "(Intercept)"] #remove the intercept










#define training (70%) and testing (30%) samples
set.seed(1234)
train = rbinom(n = nrow(mdata), size = 1, prob = 0.7) %>% as.logical()



rm(mdata)




```


\
\



# Variable Selection  

```{r vs}

#Since over-dispersion doesn't really matter until inference (calculating SEs), 
#it's fine to use Poisson for variable selection.  Just need to run the final model with NB.

#about 8 minutes
invisible(gc())
start = Sys.time()
start
cv_fit = cv.glmnet(y = Y[train], 
                   x = X[train, ],
                   weights = W[train],
                   trace.it = interactive(), #print progress
                   nfolds = 50,
                   family = "poisson", 
                   type.measure = "deviance", #use deviance as loss function
                   alpha = 1  #LASSO
                   
)
stop = Sys.time()
stop - start
beepr::beep()


saveRDS(cv_fit, "Results/charge_cv_fit.RDS")



plot(cv_fit)




# lambda.min is the value of lambda that gives minimum mean cross-validated error. 
log(cv_fit$lambda.min)
coef(cv_fit, s = "lambda.min") %>% as.matrix() %>% 
	data.frame() %>%  rownames_to_column() %>% 
	filter(X1 != 0) %>% arrange(rowname)
#way too many vars




# lambda.1se is the most regularized model with loss that is within one standard error of model with the least loss.
log(cv_fit$lambda.1se)
coef(cv_fit, s = "lambda.1se") %>% as.matrix() %>% 
	data.frame() %>%  rownames_to_column() %>% 
	filter(X1 != 0) %>% arrange(rowname)
#takes out everything except disp_ip_recode, edevent, 
# elixhauser, female, hosp_trauma_level, hr_group, 
# injury_brain, injury_cspine, injury_hollow_abd, injury_le_fx, 
# injury_lspine, injury_pelvic_fx, injury_rib_fx, injury_skull_fx, 
# injury_solid_abd, injury_tspine, injury_ue_fx, 
# los_ip, npr_ip, pDeath, year_whole, zipinc_qrtl

#takes out all ages, age_group, injury_cardio_pulm, and pay1_recode



#final choice:
coef(cv_fit, s = "lambda.1se") %>% as.matrix()

coefs = coef(cv_fit, s = "lambda.1se")
coefs = data.frame( var = rownames(coefs),  b = matrix(coefs)  )
coefs = coefs[coefs$b != 0, ]
coefs

rm(cv_fit, X, Y, W, start, stop)


```





\
\
\
\


# Testing Model

```{r testing model}

mdata = read_fst("Data/final/final combined dataset.fst",
                 columns = c(
                   
                   'totchg', 
                   
                   #vars for case mix adjustment and clustering
                   'hosp_ed', "neds_stratum", 
                   'year_whole',
                   'discwt', 
                   'high_risk', "hr_group",
                   'pDeath', 'elixhauser',
                   'hosp_trauma_level',
                   'age', 
                   
                   #vars selected by LASSO procedure
                   "disp_ip_recode", "edevent", 
                   'female', "los_ip", "npr_ip",
                   "zipinc_qrtl",
                   
                   "injury_brain", "injury_cspine", "injury_hollow_abd",
                   "injury_le_fx", "injury_lspine", "injury_pelvic_fx",
                   "injury_rib_fx", "injury_skull_fx", "injury_solid_abd",
                   "injury_tspine", "injury_ue_fx"
                   
                 )
)


mdata = mdata %>% filter(high_risk == "High-risk") %>% select(-high_risk)


test = mdata %>% filter(train != TRUE)
format(nrow(test), big.mark = ",")
format(round(sum(test$discwt)), big.mark = ",")



#Winsorize values greater than the 99th percentile
cut = quantile(mdata$totchg, probs = 0.99)
cut
test$totchg_winsor = test$totchg
test$totchg_winsor[test$totchg > cut] = cut+1
test$totchg = NULL



test$age_capped = if_else(test$age > 89, 90L, test$age)




#set reference levels for a few vars
test$edevent = factor(test$edevent) %>% 
  relevel(ref = "Admitted as inpatient")  #was "Treated and released"

test$disp_ip_recode = factor(test$disp_ip_recode) %>% 
  relevel(ref = "Not admitted as inpatient")



#if these are missing and pt was not admitted, set to zero
test$los_ip[which(is.na(test$los_ip) & test$edevent != "Admitted as inpatient" )] = 0
test$npr_ip[which(is.na(test$npr_ip) & test$edevent != "Admitted as inpatient")] = 0





test$hosp_ed = as.factor(test$hosp_ed)

test$hosp_trauma_level = factor(test$hosp_trauma_level,
                                ordered = FALSE)

test$zipinc_qrtl = factor(test$zipinc_qrtl,
                                ordered = FALSE)


test = test %>% mutate_if(is.logical, as.integer)


test$pDeath = test$pDeath*100

test$year_whole = factor(test$year_whole)




#adjust names for multiple comparisons later
test$hr_group = str_replace_all(test$hr_group, " ", "_")
test$hr_group = str_replace_all(test$hr_group, "-", "_")
test$hr_group = str_replace_all(test$hr_group, ",", "")
test$hr_group = str_replace_all(test$hr_group, "_&_", "_")
test$hr_group = str_replace_all(test$hr_group, "/", "_")



#set Pedal Cycle Related as reference group (largest)
test$hr_group = factor(test$hr_group)
test$hr_group = relevel(test$hr_group, "Pedal_Cycle_Related")






rm(mdata)


#~4 mins if using `discrete = TRUE`
start = Sys.time()
start
charge_mod = bam(totchg_winsor ~ 
								 	#level 1 fixed effects:
								 	hr_group +
								 	
								 	age_capped + female + zipinc_qrtl + 
								 	
								 	#excluding edevent since it really doesn't tell us anything los_ip doesnt
								 	#plus, it leads to issues estimating disp_ip_recode since they are non-orthogonal
								 	#edevent +  
								 	disp_ip_recode + 
								 	los_ip + npr_ip + 
								 	
								 	injury_brain + injury_cspine + injury_hollow_abd +
								 	injury_le_fx + injury_lspine + injury_pelvic_fx +
								 	injury_rib_fx + injury_skull_fx + injury_solid_abd +
								 	injury_tspine + injury_ue_fx +
								 	
								 	pDeath + elixhauser +
								 	
								 	#level 2 fixed effects:
								 	hosp_trauma_level + year_whole + 
								 	
								 	#random effects via smoothers:
								 	s(hosp_ed, bs='re') + s(neds_stratum, bs="re"),
								 
								 family = nb(link = "log"),  #estimates theta on its own :)
								 weights = test$discwt,
								 data = test,
								 
								 discrete = TRUE,
								 control = list(nthreads = 4, trace = interactive())
								 
)
end = Sys.time()
beepr::beep()

end - start

saveRDS(charge_mod, "Results/charge_mod.RDS")

options(show.signif.stars = F)


mod_sum = summary(charge_mod, re.test = F)
mod_sum$r.sq
mod_sum$dev.expl



#check residuals
ypred = predict(charge_mod, type = "link")
d_resid = residuals(charge_mod, type = 'deviance')
par(mfrow=c(1, 2))
hist(d_resid, breaks = 100)
smoothScatter(x=ypred, y=d_resid, nbin=800)
par(mfrow=c(1, 1))
rm(ypred, d_resid)



fixed = extract_fixed(charge_mod, ci_level = 1-0.005, digits = 5) 
colnames(fixed) = c("term", "value", "se", "t", "p", "lower_ci", "upper_ci")
fixed$RR = sprintf("%.2f", exp(fixed$value))
fixed$CI = paste0( sprintf("%.2f", exp(fixed$lower_ci)), ", ", 
									 sprintf("%.2f", exp(fixed$upper_ci)))
fixed$pval = if_else(fixed$p < 0.0001, "< 0.0001", sprintf("%.3f", fixed$p))
TCCD::my_print(fixed)






#pairwise comparisons
# multcomp's function to automatically make contrasts (mcp) doesn't work with models from bam().
# solution from https://stats.stackexchange.com/questions/376237/correcting-for-multiple-pairwise-comparisons-with-gam-objects-mgcv-in-r


library(multcomp)


#comparisons
#Pedal_Cycle_Related is the ref. group in the model
pairs = test %>% distinct(hr_group) %>% pull() %>% 
	sort() %>% as.character() %>% 
	combn(2) %>% t() %>% data.frame()

library(glue)


#all comparisons against Pedal_Cycle_Related
comps = pairs %>% 
	filter(X1 == "Pedal_Cycle_Related") %>% 
	glue_data("hr_group{X2} = 0")



#all other comparisons
pairs = pairs %>% 
	filter(X1 != "Pedal_Cycle_Related") %>% 
	filter(X2 != "Pedal_Cycle_Related")


comps = c(comps, 
					pairs %>% glue_data("hr_group{X1} - hr_group{X2} = 0")
)




pw_comps = glht(charge_mod, 
								linfct = comps
								)



summary(pw_comps, test = adjusted("holm") )


res = confint(pw_comps, level = 1-0.005, calpha = univariate_calpha())$confint

res = res %>% as.data.frame() %>% rownames_to_column()

res$ps = summary(pw_comps, test = adjusted("none") )$test$pvalues
res$ps_adj = summary(pw_comps, test = adjusted("holm") )$test$pvalues

res$RR = scales::comma(exp(res$Estimate), accuracy = 0.01)
res$ci_l = scales::comma(exp(res$lwr), accuracy = 0.01)
res$ci_u = scales::comma(exp(res$upr), accuracy = 0.01)

res$RR_CI = paste0(res$RR, " (", res$ci_l, " - ", res$ci_u, ")")

res$ps = scales::pvalue(res$ps, accuracy = 0.0001)
res$ps_adj = scales::pvalue(res$ps_adj, accuracy = 0.0001)


res %>% dplyr::select(rowname, RR:RR_CI, ps, ps_adj) %>% TCCD::my_print()






```


\
\
\
\





```{r} 
pander::pander(sessionInfo())
```
